แก้ไฟล์งานไปในโฟเดอร์กลุ่มเรียบร้อย
