+++ MMK22/src/Profile.js
@@ -1,0 +1,24 @@
+import {useState} from "react";
+import './Profile.css';
+import { Avatar } from './components/Avatar.js';
+
+function Profile() {
+  const [name, setName] = useState("Unknown");
+
+
+  return (
+    <div className="Profile">
+      <Avatar name={name}/>
+      <label className="label1" for="displayname"/>
+      <input 
+        id="displayname"
+        type="text" 
+        className="text" 
+        value={name}
+        onChange={e => setName(e.target.value)}
+      />
+    </div>
+  );
+}
+
+export default Profile;
