+++ MMK22/src/components/Userlist.js
@@ -1,8 +1,6 @@
-function Userlist () {
+export function Userlist () {
   return (
     <div className="userlist">
     </div>
   )
 }
-
-export default Userlist;
