+++ MMK22/src/css/Avatar.css
@@ -1,7 +1,8 @@
 .avatar {
+  background-color: rgb(228, 195, 236);
   width: 100px;
   height: 100px;
   border-radius: 50%;
-  border: solid 1px gray;
+  border: solid 1px rgb(5, 5, 5);
   overflow: hidden;
 }
