+++ MMK22/src/index.js
@@ -1,8 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import Router from './Router';
 import reportWebVitals from './reportWebVitals';
+import Router from './Router.js';
+
+// index.html --> index.js --> Router.js --> App.js
 
 ReactDOM.render(
   <React.StrictMode>
