--- src/Profile.js
@@ -1,27 +1,0 @@
-import App from './components/App.js';
-import {useState} from "react";
-import './Profile.css';
-import Avatar from './components/Avatar.js';
-
-function Profile() {
-  const [name, setName] = useState("Unknown");
-
-
-  return (
-    <App>
-      <div className="Profile">
-        <Avatar name={name}/>
-        <label className="label1" htmlFor="displayname"/>
-        <input 
-          id="displayname"
-          type="text" 
-          className="text" 
-          value={name}
-          onChange={e => setName(e.target.value)}
-        />
-      </div>
-    </App>
-  );
-}
-
-export default Profile;
